refactor: wrap table1 in class
